# api/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger Photo Caption Contest - OpenAPI 3.0
  description: Swagger Photo Caption Contest - OpenAPI 3.0
  contact:
    name: Derek R. Cox
    url: www.thecloudguy.org
    email: derek.ryan.cox+swagger@thecloudguy.org
  license:
    name: GNU GENERAL PUBLIC LICENSE
    url: https://www.gnu.org/licenses/
  version: 1.0.0
...
servers:
  - url: http://localhost:3000
    variables: {}
    description: Localhost
...
paths:
...
  /route/users/:
    summary: Users Root
    description: Root of the users URL
    get:
      tags:
        - /route/users
      summary: Get List of Users
      description: Get List of Users
      operationId: getUsers
      responses:
        '200':
          content:
            application/json:
              examples:
                GetUsers:
                  value: |-
                    {
                        "status": "Success",
                        "message": "User Information retrieved",
                        "data": [
                            {
                                "id": 647,
                                "name": "User_Test",
                                "email": "user.test@testdomain.com",
                                "password": "P@ssw0rd",
                                "createdAt": "2023-03-26T00:10:42.949Z",
                                "updatedAt": "2023-03-26T00:10:42.949Z"
                            }
                        ]
                    }
          description: success
        default:
          description: Default error sample response
      x-swagger-router-controller: Routesusers
    post:
      tags:
        - /route/users
      summary: Add User
      description: Add User
      operationId: addUser
      responses:
        '201':
          content:
            application/json:
              examples:
                PostUsers:
                  value: |-
                    {
                        "status": "Success",
                        "message": "New User Created",
                        "data": {
                            "id": 647,
                            "name": "User_Test",
                            "email": "user.test@testdomain.com",
                            "password": "P@ssw0rd",
                            "updatedAt": "2023-03-26T00:10:42.949Z",
                            "createdAt": "2023-03-26T00:10:42.949Z"
                        }
                    }
          description: success
        default:
          description: Default error sample response
      x-swagger-router-controller: Routesusers
...